improve multi-monitor configuration
Using 'output * <cmd>' fails if outputs are set inactive (for instance,
a closed laptop screen). Created sway-get-active-outputs (which requires
`jq`) to get a list of all *active* outputs and use that instead of *.
Also managed to get the lid switch working; however, the initial startup
state is still wrong (output enabled) if the lid is closed.